# openapi/proxy-openapi.yaml
openapi: 3.1.0
info:
  title: Pulse Alpaca Compatibility Proxy
  version: 0.1.0
  description: >-
    Initial contract for first-party authentication and Alpaca-compatible routes.
    Alpaca-facing schemas are completed route-by-route from current official references.
servers:
  - url: https://proxy.example.com
paths:
  /auth/login:
    post:
      summary: Authenticate with a username and password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, password]
              properties:
                username: {type: string}
                password: {type: string, format: password}
      responses:
        '200':
          description: Bearer token issued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LoginResponse'
        '401': {$ref: '#/components/responses/AuthInvalid'}
        '429': {description: Login rate limited}
...
  /v2/account:
    get:
      summary: Alpaca-compatible account endpoint
      description: >-
        Forwards GET /v2/account to the configured Alpaca paper or live API
        with fixed server-side credentials. The caller must supply a proxy token
        with the trading:read scope.
      security: [{bearerAuth: []}]
      responses:
        '200': {description: Alpaca account response; schema to be synchronized from official reference}
        '401': {$ref: '#/components/responses/AuthInvalid'}
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque proxy token
  schemas:
    LoginResponse:
      type: object
      required: [access_token, token_type, expires_in, user]
      properties:
        access_token: {type: string}
        token_type: {type: string, example: Bearer}
        expires_in: {type: integer, example: 3600}
        user:
          type: object
          required: [id, username]
          properties:
            id: {type: string}
            username: {type: string}
  responses:
    AuthInvalid:
      description: Missing, invalid, expired, or revoked proxy token